Keep explorer integration within quality budgets
