+++ internal/platform/architecture/architecture_test.go
@@ -818,16 +818,6 @@ func TestAnalyticsGeneratedAPIIsCapabilityOwned(t *testing.T) {
 	}
 }
 
-func TestAnalyticsGeneratedAPIUsesExplorationOwnedVisualizationContracts(t *testing.T) {
-	body, err := os.ReadFile(filepath.Join(repoRoot(t), "internal", "analytics", "api", "gen", "request_models.gen.go"))
-	if err != nil {
-		t.Fatalf("read generated Analytics request models: %v (run task api:generate first)", err)
-	}
-	if strings.Contains(string(body), "internal/dashboard/visualization/ir") {
-		t.Fatal("generated Analytics request models import dashboard visualization IR; authored exploration formats must remain local")
-	}
-}
-
 func TestProjectGeneratedAPIIsCapabilityOwned(t *testing.T) {
 	rule, ok := ClassifyPackage("internal/project/api/gen")
 	if !ok {
